skip some blocker tests that are caused by libvmware-vix.so
Skip two tests that end badly for which I could not find a solution.
a) one test gets python into a segmentation fault and stops the whole
test suite
b) one test makes an invalid authentication call to check whether
vmware responds with an error (and it does as it should), but
afterwards we cannot make valid authentication calls from pyvix.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@gmail.com>